Duration have been transferred in the separate component
