+++ src/components/CreateCourse/CreateCourse.jsx
@@ -6,11 +6,10 @@ import { useState } from 'react';
 import { Authors } from './components/Authors/Authors';
 import { CourseAuthors } from './components/CourseAuthors/CourseAuthors';
 import { AddAuthor } from './components/AddAuthor/AddAuthor';
+import { Duration } from './components/Duration/Duration';
 
 export const CreateCourse = ({ authorsList, createNewAuthor }) => {
 	const [selectedAuthors, setSelectedAuthors] = useState([]);
-	const [min, setMinutes] = useState(0);
-
 	const addAuthor = (id) => {
 		setSelectedAuthors([...selectedAuthors, id]);
 	};
@@ -18,13 +17,6 @@ export const CreateCourse = ({ authorsList, createNewAuthor }) => {
 	const deleteAuthor = (id) => {
 		setSelectedAuthors(selectedAuthors.filter((author) => author !== id));
 	};
-
-	const handleInputChange = ({ target: { value } }) => {
-		setMinutes(value);
-	};
-	function padTo2Digits(num) {
-		return num.toString().padStart(2, '0');
-	}
 
 	return (
 		<div className={styles.main}>
@@ -36,23 +28,7 @@ export const CreateCourse = ({ authorsList, createNewAuthor }) => {
 			<div className={styles.row}>
 				<div className={styles.column}>
 					<AddAuthor createNewAuthor={createNewAuthor} />
-					<div className={styles.durationDiv}>
-						<h3 className={styles.h3}>Duration</h3>
-						<Input
-							min='0'
-							type='number'
-							labelText='Duration'
-							placeholderText='Enter duration in minutes'
-							onChange={handleInputChange}
-						/>
-						<div className={styles.durationCounter}>
-							Duration:{' '}
-							<span className='typography--bold typography--extra-large'>
-								{padTo2Digits(Math.floor(min / 60))} : {padTo2Digits(min % 60)}
-							</span>{' '}
-							hours
-						</div>
-					</div>
+					<Duration />
 				</div>
 				<div className={styles.column}>
 					<Authors
